01 in 09 done

+++ C++ module 09/ex01/RPN.hpp
@@ -1,10 +1,11 @@
 #pragma once
 #include <iostream>
+#include <sstream>
 #include <stack>
 
 class RPN {
     private:
-        std::stack<double> stk;
+        std::stack<int> stk;
     public:
         RPN();
         RPN(const RPN &other);
